more changes

+++ mcp/command_struct.h
@@ -74,10 +74,10 @@ struct CommandDataStruct {
   unsigned short int sucks;
   unsigned short int lat_range; 
 
-  short int apcu_reg;
+  double apcu_reg;
   double  apcu_trim;
   short int apcu_auto;
-  short int dpcu_reg;
+  double dpcu_reg;
   double dpcu_trim;
   short int dpcu_auto;
